+++ lua/obsidian/commands/extract_note.lua
@@ -35,7 +35,6 @@ return function(client, data)
   client:update_ui(0)
 
   -- add the selected text to the end of the new note
-  local open_in = util.get_open_strategy(client.opts.open_notes_in)
-  vim.cmd(open_in .. tostring(note.path))
+  client:open_note(note)
   vim.api.nvim_buf_set_lines(0, -1, -1, false, content)
 end