+++ lua/obsidian/commands/new.lua
@@ -1,10 +1,7 @@
-local util = require "obsidian.util"
-
 ---@param client obsidian.Client
 return function(client, data)
   ---@type obsidian.Note
   local note
-  local open_in = util.get_open_strategy(client.opts.open_notes_in)
   if data.args:len() > 0 then
     note = client:new_note(data.args)
   else
@@ -16,5 +13,5 @@ return function(client, data)
     end
     note = client:new_note(title)
   end
-  vim.api.nvim_command(open_in .. tostring(note.path))
+  client:open_note(note)
 end